Display add project dialog on '+' click
